implicit conversion broke tests. I have remaved them and will need to figure out how to add them back in later. The explicit as_bool() as_long() and as_double() methods are a good addition anyway.
Start to write a better regression test suite. THe existing tests are weak.


git-svn-id: https://cxx.svn.sourceforge.net/svnroot/cxx/trunk@119 cb2f5747-f50d-0410-b093-b27f9c363ba5

+++ CXX/Demo/pycxx_iter.hxx
@@ -73,7 +73,7 @@ public:
         else
             return NULL;    // signal StopIteration
 
-        Py::Int Result(ct);
+        Py::Long Result(ct);
         Result.increment_reference_count();
         return Result.ptr();
     }
